Edited several lines for clarity
Edited several lines of 'text only' to make it more understandable.
----------------------------------------------------
There are still a couple of problems:
----------------------------------------------------
Line 2 [title}:  Is not descriptive of the page. The subject of page is about "How MyEtherWallet Can Create (a) Wallet(s) Offline?" VS. "Send Offline" as it doesn't give any example of the sending transaction at all, broadcasting yes but not enough. The word SEND is only on the page one time; in the title.
Line 13: ... every address already "exists"
[conflicts with] 
Line 15: ... randomly generates both a private key and an address (public key)