Update adding-new-token-and-sending-custom-tokens.md
"This will remove it from the interface." is not detailed enough : is the added or removed token is in the browser? Where is it stored? in a cookie? If I change my computer or login on another computer none of my added token will show up. I can't find any information on Internet. Can you explain that point?
Thanks
regards